updaterate for codum.token arg fix, deletetx for cotumpresale (#27)
* fixed missing closure on buycodum contribution_table.emplace();

* added .vscode to .gitignore;

* codum-core#5:
* rolled back rate to int64_t due to backw.compatibility;
* removed some stale code;

* codum-core#5:
- moved start 4 days back for testing, correct start at EOL;

* returned start to original date; optimized some constant deftypes;

* removed .vscode;

* * token updaterate takes argument as uint32_t;
* added deletetx action to remove unvalidated transactions;

+++ contracts/codum.token/codum.token.cpp
@@ -156,7 +156,7 @@ void token::distribcontr(account_name from, account_name to, asset quantity, str
     // transfer(from, to, quantity, memo); // needs to be inlined.
 }
 
-void token::updaterate(uint8_t network, uint64_t rate) {
+void token::updaterate(uint8_t network, uint32_t rate) {
     require_auth(_self);
     exrates exrates_table(_self, _self); // code: _self, scope: _self
     auto itr = exrates_table.find(network);
